+++ package.json
@@ -68,7 +68,6 @@
     "react": "^16.0.0",
     "react-dom": "^16.0.0",
     "selection-position": "^1.0.0",
-    "showdown": "^1.0.0",
     "string": "^3.0.0",
     "styled-components": "^4.0.0",
     "underscore": "^1.0.0",
