+++ packages/markdown/__tests__/index.js
@@ -1,0 +1,6 @@
+import { converter } from '../src'
+
+test("Converter doesn't throw", () => {
+  const html = converter.makeHtml('# Heading')
+  expect(html).toEqual('<h1 id="heading">Heading</h1>')
+})
