+++ packages/markdown/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@serlo/markdown",
+  "version": "0.0.0",
+  "main": "src/index.js",
+  "dependencies": {
+    "showdown": "^1.0.0"
+  },
+  "private": true
+}
